refactor: catch errors before the request cycle starts
Previously we did not catch exceptions during the creation of a request. We also called `before_request` before the request creation succeeds. Flask's `before_request` provides the handlers with a request object so we do the same thing now.

If the request was never created we don't call `after_request` either