refactor: request lifecycle and handling
Streamline the way requests work and make sure that cancellation can happen in a predictable fashion. We now ensure streaming responses and streaming requests are cancelled correctly.

We also have moved Response so that it must be explicitly requested by handlers. This makes
it a bit clearer than using `request.response`